splits out progress bar into separate component
