+++ react-front/public/components/ConfigChange/ConfigChangeStep2.js
@@ -1,5 +1,6 @@
 import React from "react";
-import getData from "../../utils/getData";
+// import getData from "../../utils/getData";
+import ConfigChangeProgressBar from "./ConfigChangeProgressBar";
 import { postData } from "../../utils/sendData";
 //WORK IN PROGRESS
 
@@ -9,7 +10,7 @@ class ConfigChangeStep2 extends React.Component {
     deviceSync: [],
     deviceSyncStatus: [],
     deviceSyncJobId: [],
-    jobsData: [],
+    // jobsData: [],
     jobStartTime: [],
     jobFinishTime: []
     // finishedDevices: [],
@@ -32,13 +33,14 @@ class ConfigChangeStep2 extends React.Component {
             deviceSync: data.data,
             deviceSyncStatus: data.status,
             deviceSyncJobId: data.job_id
+            // startJobIdSync: data.job_id
             // jobStartTime: data.start_time,
             // jobFinishTime: data.finish_time,
             // exception: data.exception
           },
-          () => {
-            this.syncStatus();
-          },
+          // () => {
+          //   this.syncStatus();
+          // },
           () => {
             console.log("this is new state", this.state.deviceSync);
           }
@@ -47,69 +49,74 @@ class ConfigChangeStep2 extends React.Component {
     });
   };
 
-  syncStatus = () => {
-    // let jobId = this.state.deviceSyncJobId;
-    let jobId = "5ddbe1548b2d390c963b97d8";
-    const credentials =
-      "eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiJ9.eyJpYXQiOjE1NzEwNTk2MTgsIm5iZiI6MTU3MTA1OTYxOCwianRpIjoiNTQ2MDk2YTUtZTNmOS00NzFlLWE2NTctZWFlYTZkNzA4NmVhIiwic3ViIjoiYWRtaW4iLCJmcmVzaCI6ZmFsc2UsInR5cGUiOiJhY2Nlc3MifQ.Sfffg9oZg_Kmoq7Oe8IoTcbuagpP6nuUXOQzqJpgDfqDq_GM_4zGzt7XxByD4G0q8g4gZGHQnV14TpDer2hJXw";
+  // syncStatus = () => {
+  //   // let jobId = this.state.deviceSyncJobId;
+  //   let jobId = "5ddbe1548b2d390c963b97d8";
+  //   const credentials =
+  //     "eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiJ9.eyJpYXQiOjE1NzEwNTk2MTgsIm5iZiI6MTU3MTA1OTYxOCwianRpIjoiNTQ2MDk2YTUtZTNmOS00NzFlLWE2NTctZWFlYTZkNzA4NmVhIiwic3ViIjoiYWRtaW4iLCJmcmVzaCI6ZmFsc2UsInR5cGUiOiJhY2Nlc3MifQ.Sfffg9oZg_Kmoq7Oe8IoTcbuagpP6nuUXOQzqJpgDfqDq_GM_4zGzt7XxByD4G0q8g4gZGHQnV14TpDer2hJXw";
 
-    console.log("this API call is automatic");
-    let url = `https://tug-lab.cnaas.sunet.se:8443/api/v1.0/job/${jobId}`;
-    getData(url, credentials).then(data => {
-      console.log("this should be data.data.jobs", data.data.jobs);
-      {
-        this.setState(
-          {
-            jobsData: data.data.jobs
-          },
-          () => {
-            console.log("this is jobs data", this.state.jobsData);
-          }
-        );
-      }
-    });
-  };
+  //   console.log("this API call is automatic");
+  //   let url = `https://tug-lab.cnaas.sunet.se:8443/api/v1.0/job/${jobId}`;
+  //   getData(url, credentials).then(data => {
+  //     console.log("this should be data.data.jobs", data.data.jobs);
+  //     {
+  //       this.setState(
+  //         {
+  //           jobsData: data.data.jobs
+  //         },
+  //         () => {
+  //           console.log("this is jobs data", this.state.jobsData);
+  //         }
+  //       );
+  //     }
+  //   });
+  // };
 
   render() {
     // console.log("these are props (in Workflow)", this.props);
     let syncMessage = this.state.deviceSync;
     let syncStatus = this.state.deviceSyncStatus;
     let syncJobId = this.state.deviceSyncJobId;
-    let jobsData = this.state.jobsData;
+    let jobStartTime = "today, 09:34";
+    let jobFinishTime = "today, 09:44";
+    // let exceptionText = job.exception;
 
-    function randomIntFromInterval(min, max) {
-      return Math.floor(Math.random() * (max - min + 1) + min);
-    }
+    console.log("this is job_id send from parent to progressbar:", syncJobId);
+    // let jobsData = this.state.jobsData;
 
-    let jobsProgress = jobsData.map((job, i) => {
-      // let totalDevices = job.result._totals.selected_devices;
-      // let jobStatus = job.status;
-      // let finishedDevices = job.finished_devices.length;
-      let jobStartTime = job.start_time;
-      let jobFinishTime = job.finish_time;
-      let exceptionText = job.exception;
-      
-      let finishedDevices = randomIntFromInterval(0, 100);
-      console.log("this is finsihedDevices", finishedDevices);
-      let totalDevices = 100;
-      return (
-        <div id="progressbar">
-          <p>Progress bar</p>
-          <progress
-            min="0"
-            max={totalDevices}
-            value={finishedDevices}
-          ></progress>
-          <label>
-            {finishedDevices}/{totalDevices} devices finished
-          </label>
-          <p>Other job info</p>
-          <p>status: {syncStatus}</p>
-          <p>start time: {jobStartTime}</p>
-          <p>finish time: {jobFinishTime}</p>
-        </div>
-      );
-    });
+    // function randomIntFromInterval(min, max) {
+    //   return Math.floor(Math.random() * (max - min + 1) + min);
+    // }
+
+    // let jobsProgress = jobsData.map((job, i) => {
+    //   // let totalDevices = job.result._totals.selected_devices;
+    //   // let jobStatus = job.status;
+    //   // let finishedDevices = job.finished_devices.length;
+    //   let jobStartTime = job.start_time;
+    //   let jobFinishTime = job.finish_time;
+    //   let exceptionText = job.exception;
+
+    //   let finishedDevices = randomIntFromInterval(0, 100);
+    //   console.log("this is finsihedDevices", finishedDevices);
+    //   let totalDevices = 100;
+    //   return (
+    //     <div id="progressbar">
+    //       <p>Progress bar</p>
+    //       <progress
+    //         min="0"
+    //         max={totalDevices}
+    //         value={finishedDevices}
+    //       ></progress>
+    //       <label>
+    //         {finishedDevices}/{totalDevices} devices finished
+    //       </label>
+    //       <p>Other job info</p>
+    //       <p>status: {syncStatus}</p>
+    //       <p>start time: {jobStartTime}</p>
+    //       <p>finish time: {jobFinishTime}</p>
+    //     </div>
+    //   );
+    // });
     return (
       <div className="workflow-container">
         <div className="workflow-container__header">
@@ -131,9 +138,14 @@ class ConfigChangeStep2 extends React.Component {
             <p>{syncStatus}</p>
             <p>{syncJobId}</p>
           </div>
-          <div>          
-            {jobsProgress}
+          <div>
+            <ConfigChangeProgressBar jobId={this.state.deviceSyncJobId} />
+            {/* {jobsProgress} */}
           </div>
+          <p>Other job info</p>
+          <p>status: {syncStatus}</p>
+          <p>start time: {jobStartTime}</p>
+          <p>finish time: {jobFinishTime}</p>
         </div>
       </div>
     );
